Move x64 LUT to the x64 recompiler name space
